Switched to using new usgsStreamgagesLayer
