+++ app/client/src/components/pages/Community.Tabs.Overview.js
@@ -32,6 +32,7 @@ import { tabsStyles } from 'components/shared/ContentTabs';
 import VirtualizedList from 'components/shared/VirtualizedList';
 // contexts
 import { useFetchedDataState } from 'contexts/FetchedData';
+import { useLayers } from 'contexts/Layers';
 import { LocationSearchContext } from 'contexts/locationSearch';
 import { useServicesContext } from 'contexts/LookupFiles';
 // utilities
@@ -121,8 +122,9 @@ function Overview() {
     watershed,
     visibleLayers,
     setVisibleLayers,
-    usgsStreamgagesLayer,
   } = useContext(LocationSearchContext);
+
+  const { usgsStreamgagesLayer } = useLayers();
 
   const [waterbodiesDisplayed, setWaterbodiesDisplayed] = useState(true);
 
@@ -518,8 +520,9 @@ function MonitoringAndSensorsTab({
     monitoringLocations,
     monitoringLocationsLayer,
     watershed,
-    usgsStreamgagesLayer,
   } = useContext(LocationSearchContext);
+
+  const { usgsStreamgagesLayer } = useLayers();
 
   const services = useServicesContext();
